move GateSearchState to separate file
